+++ jimp2_projekt/src/key.h
@@ -6,11 +6,11 @@
 
 typedef struct key_type
 {
-	int value;
+	long int value;
 	int length;
 } key_type;
 
-void AssignKeys(node_t head, key_type *keys, int val, int len);
+void AssignKeys(node_t head, key_type *keys, long int val, int len);
 
 char* KeyToCode(key_type key);
